Refactor code and remove unnecessary things
