+++ src/main/java/nl/mpdev/backend_spring_boot_tech_it_easy_controller/controllers/TelevisionsController.java
@@ -4,7 +4,6 @@ import jakarta.validation.Valid;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.general.IdInputDto;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.complete.TelevisionCompleteInputDto;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.complete.TelevisionCompleteOutputDTO;
-import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.sales.TelevisionSalesInputDto;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.sales.TelevisionSalesOutputDto;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.services.TelevisionService;
 import org.springframework.http.HttpStatus;
@@ -18,8 +17,6 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/v1/televisions")
 public class TelevisionsController {
-
-//  List<Television> televisionDataBase = new ArrayList<>();
 
     private final TelevisionService televisionService;
 
@@ -57,25 +54,10 @@ public class TelevisionsController {
         return ResponseEntity.created(uri).body(newTelevision);
     }
 
-    @PostMapping("/sales")
-    public ResponseEntity<TelevisionSalesOutputDto> addTelevisionSales(@Valid @RequestBody TelevisionSalesInputDto televisionSalesInputDto) {
-        TelevisionSalesOutputDto newTelevision = televisionService.addTelevisionSales(televisionSalesInputDto);
-        // I try to figure out how i can give back the URI more dynamically for this case but I dont know how to do it yet.
-        URI uri = URI.create(ServletUriComponentsBuilder.fromCurrentContextPath().path("/api/v1/televisions/" + newTelevision.getId() + "/sales").toUriString());
-        return ResponseEntity.created(uri).body(newTelevision);
-    }
-
-
-
     // PUT
     @PutMapping("/{id}")
     public ResponseEntity<TelevisionCompleteOutputDTO> updateTelevision(@PathVariable Long id, @Valid @RequestBody TelevisionCompleteInputDto televisionCompleteInputDto) {
         return ResponseEntity.ok().body(televisionService.updateTelevision(id, televisionCompleteInputDto));
-    }
-
-    @PutMapping("/{id}/sales")
-    public ResponseEntity<TelevisionSalesOutputDto> updateTelevisionSales(@PathVariable Long id, @Valid @RequestBody TelevisionSalesInputDto televisionSalesInputDto) {
-        return ResponseEntity.ok().body(televisionService.updateTelevisionSales(id, televisionSalesInputDto));
     }
 
     @PutMapping("/{id}/remotes")
@@ -98,12 +80,6 @@ public class TelevisionsController {
     public ResponseEntity<TelevisionCompleteOutputDTO> updateTelevisionFields(@PathVariable Long id, @Valid @RequestBody TelevisionCompleteInputDto televisionCompleteInputDto) {
         return ResponseEntity.ok().body(televisionService.updateTelevisionFields(id, televisionCompleteInputDto));
     }
-
-    @PatchMapping("/{id}/sales")
-    public ResponseEntity<TelevisionSalesOutputDto> updateTelevisionSalesFields(@PathVariable Long id, @Valid @RequestBody TelevisionSalesInputDto televisionSalesInputDto) {
-        return ResponseEntity.ok().body(televisionService.updateTelevisionSalesFields(id, televisionSalesInputDto));
-    }
-
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
