+++ src/main/java/nl/mpdev/backend_spring_boot_tech_it_easy_controller/dtos/cimodules/CIModuleCompleteInputDto.java
@@ -1,8 +1,10 @@
 package nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.cimodules;
 
-import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 public class CIModuleCompleteInputDto {
   private Long id;
   private String name;
