+++ src/main/java/nl/mpdev/backend_spring_boot_tech_it_easy_controller/dtos/cimodules/CIModuleCompleteOutputDto.java
@@ -1,11 +1,9 @@
 package nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.cimodules;
 
-import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
-@Data
-public class CIModuleCompleteOutputDto {
-  private Long id;
-  private String name;
-  private String type;
-  private Double price;
+@Getter
+@Setter
+public class CIModuleCompleteOutputDto extends CIModuleCompleteInputDto {
 }
