+++ src/main/java/nl/mpdev/backend_spring_boot_tech_it_easy_controller/dtos/remotes/RemoteCompleteInputDto.java
@@ -1,8 +1,10 @@
 package nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.remotes;
 
-import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 public class RemoteCompleteInputDto {
   private Long id;
   private String compatibleWith;
