+++ src/main/java/nl/mpdev/backend_spring_boot_tech_it_easy_controller/dtos/remotes/RemoteCompleteMapper.java
@@ -9,7 +9,7 @@ public class RemoteCompleteMapper {
 
   public Remote toEntity(RemoteCompleteInputDto remoteCompleteInputDto) {
     Remote remote = new Remote();
-    BeanUtils.copyProperties(remote, remoteCompleteInputDto);
+    BeanUtils.copyProperties(remoteCompleteInputDto, remote);
     remote.setId(null);
     return remote;
   }
