+++ src/main/java/nl/mpdev/backend_spring_boot_tech_it_easy_controller/dtos/remotes/RemoteCompleteOutputDto.java
@@ -1,7 +1,9 @@
 package nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.remotes;
 
-import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
-@Data
+@Getter
+@Setter
 public class RemoteCompleteOutputDto extends RemoteCompleteInputDto {
 }
