+++ src/main/java/nl/mpdev/backend_spring_boot_tech_it_easy_controller/dtos/televisions/complete/TelevisionCompleteOutputDTO.java
@@ -4,7 +4,6 @@ import lombok.Getter;
 import lombok.Setter;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.cimodules.CIModuleCompleteOutputDto;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.remotes.RemoteCompleteOutputDto;
-import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.sales.TelevisionSalesInputDto;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.wallbrackets.WallbracketCompleteOutputDto;
 
 import java.time.LocalDate;
