+++ src/main/java/nl/mpdev/backend_spring_boot_tech_it_easy_controller/services/TelevisionService.java
@@ -2,10 +2,9 @@ package nl.mpdev.backend_spring_boot_tech_it_easy_controller.services;
 
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.general.IdInputDto;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.complete.TelevisionCompleteInputDto;
-import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.complete.TelevisionCompleteTelevisionMapper;
+import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.complete.TelevisionCompleteMapper;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.complete.TelevisionCompleteOutputDTO;
-import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.sales.TelevisionSalesInputDto;
-import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.sales.TelevisionSalesTelevisionMapper;
+import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.sales.TelevisionSalesMapper;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.dtos.televisions.sales.TelevisionSalesOutputDto;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.exceptions.GeneralException;
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.exceptions.RecordNotFoundException;
@@ -19,7 +18,6 @@ import nl.mpdev.backend_spring_boot_tech_it_easy_controller.repositories.Televis
 import nl.mpdev.backend_spring_boot_tech_it_easy_controller.repositories.WallBracketRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
-
 import java.lang.reflect.Field;
 import java.util.*;
 import java.util.stream.Collectors;
@@ -31,12 +29,12 @@ public class TelevisionService {
   private final RemoteRepository remoteRepository;
   private final CIModuleRepository ciModuleRepository;
   private final WallBracketRepository wallBracketRepository;
-  private final TelevisionCompleteTelevisionMapper televisionCompleteMapper;
-  private final TelevisionSalesTelevisionMapper televisionSalesMapper;
+  private final TelevisionCompleteMapper televisionCompleteMapper;
+  private final TelevisionSalesMapper televisionSalesMapper;
 
   public TelevisionService(TelevisionRepository televisionRepository, RemoteRepository remoteRepository,
-                           CIModuleRepository ciModuleRepository, WallBracketRepository wallBracketRepository, TelevisionCompleteTelevisionMapper televisionCompleteMapper,
-                           TelevisionSalesTelevisionMapper televisionSalesMapper) {
+                           CIModuleRepository ciModuleRepository, WallBracketRepository wallBracketRepository, TelevisionCompleteMapper televisionCompleteMapper,
+                           TelevisionSalesMapper televisionSalesMapper) {
 
     this.televisionRepository = televisionRepository;
     this.remoteRepository = remoteRepository;
@@ -45,6 +43,8 @@ public class TelevisionService {
     this.televisionCompleteMapper = televisionCompleteMapper;
     this.televisionSalesMapper = televisionSalesMapper;
   }
+
+  // GET
 
   public TelevisionCompleteOutputDTO getTelevision(Long id) {
     Television television = televisionRepository.findById(id).orElseThrow(() -> new RecordNotFoundException("Record not found"));
@@ -71,23 +71,12 @@ public class TelevisionService {
     return televisionCompleteMapper.toDto(television);
   }
 
-  public TelevisionSalesOutputDto addTelevisionSales(TelevisionSalesInputDto televisionSalesInputDto) {
-    Television television = televisionRepository.save(televisionSalesMapper.toEntity(televisionSalesInputDto));
-    return televisionSalesMapper.toDto(television);
-  }
-
   // PUT
 
   public TelevisionCompleteOutputDTO updateTelevision(Long id, TelevisionCompleteInputDto televisionCompleteInputDto) {
     Television existingTelevision = televisionRepository.findById(id).orElseThrow(() -> new RecordNotFoundException("Record not found"));
     BeanUtils.copyProperties(televisionCompleteInputDto, existingTelevision, "id");
     return televisionCompleteMapper.toDto(televisionRepository.save(existingTelevision));
-  }
-
-  public TelevisionSalesOutputDto updateTelevisionSales(Long id, TelevisionSalesInputDto televisionSalesInputDto) {
-    Television existingTelevision = televisionRepository.findById(id).orElseThrow(() -> new RecordNotFoundException("Record not found"));
-    BeanUtils.copyProperties(televisionSalesInputDto, existingTelevision, "id");
-    return televisionSalesMapper.toDto(televisionRepository.save(existingTelevision));
   }
 
   public TelevisionCompleteOutputDTO updateTelevisionWithRemote(Long id, IdInputDto idInputDto) {
@@ -124,12 +113,6 @@ public class TelevisionService {
     return televisionCompleteMapper.toDto(televisionRepository.save(existingTelevision));
   }
 
-  public TelevisionSalesOutputDto updateTelevisionSalesFields(Long id, TelevisionSalesInputDto televisionSalesInputDto) {
-    Television existingTelevision = televisionRepository.findById(id).orElseThrow(() -> new RecordNotFoundException("Record not found"));
-    setFieldsIfNotNUll(existingTelevision, televisionSalesInputDto);
-    return televisionSalesMapper.toDto(televisionRepository.save(existingTelevision));
-  }
-
   public void deleteTelevision(Long id) {
     Television existingTelevision = televisionRepository.findById(id).orElseThrow(() -> new RecordNotFoundException("Record not found"));
     televisionRepository.deleteById(id);
@@ -159,5 +142,4 @@ public class TelevisionService {
       }
     }
   }
-
 }
